refactor: move `uhyve` out of `syscall/interfaces`
