--- src/syscalls/interfaces/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod uhyve;
